ConversationalChatBot_Using_LangChain/requirements.txt
openai==1.12.0
streamlit==1.31.0
langchain==0.1.6
python-dotenv==1.0.1 # helps to create and upload the variable with respect to our applications
